# doc/gen-res-by-zupportL5/static/public/ZupportL5-api-swagger-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ZupportL5 API Documentation",
    "description": "Comprehensive OpenAPI-compliant JSON documentation for ZupportL5-API",
    "version": "1.0.0",
    "contact": {
      "email": "arcbrth@gmail.com",
      "name": "Development"
    },
    "license": {
      "url": "https://github.com/abritoh/zupportl5-api",
      "name": "Apache 2.0"
    },
    "origin": "This documentation was generated at 2025-02-14T16:17:34Z using ZupportL5 json-api-processor"
  },
  "paths": {
...
    "/api/batch/process-xmlfiles": {
      "post": {
        "summary": "processXmlFiles",
        "operationId": "processXmlFiles",
        "parameters": [
          {
            "name": "files",
            "in": "query",
            "required": false,
            "schema": {
              "type": "multipartfile[]"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "summary": "getViewForProcessXml",
        "operationId": "getViewForProcessXml",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/documents/id/{idSource}": {
      "get": {
        "summary": "getDocument",
        "operationId": "getDocument",
        "parameters": [
          {
            "name": "idSource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/documents/search": {
      "get": {
        "summary": "searchDocuments",
        "operationId": "searchDocuments",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/api/documents/all": {
      "get": {
        "summary": "getAllDocuments",
        "operationId": "getAllDocuments",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/api/dropbox/oauth2/callback": {
      "get": {
        "summary": "handleCallback",
        "operationId": "handleCallback",
        "parameters": [
          {
            "name": "",
            "in": "query",
            "required": false,
            "schema": {
              "type": "map"
            }
          },
          {
            "required": false,
            "schema": {
              "type": "httpservletrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  }
}